+++ types/basetypes/number_value_test.go
@@ -15,69 +15,6 @@ func numberComparer(i, j *big.Float) bool {
 	return (i == nil && j == nil) || (i != nil && j != nil && i.Cmp(j) == 0)
 }
 
-func TestNumberTypeValueFromTerraform(t *testing.T) {
-	t.Parallel()
-
-	type testCase struct {
-		input       tftypes.Value
-		expectation attr.Value
-		expectedErr string
-	}
-	tests := map[string]testCase{
-		"value": {
-			input:       tftypes.NewValue(tftypes.Number, 123),
-			expectation: NewNumberValue(big.NewFloat(123)),
-		},
-		"unknown": {
-			input:       tftypes.NewValue(tftypes.Number, tftypes.UnknownValue),
-			expectation: NewNumberUnknown(),
-		},
-		"null": {
-			input:       tftypes.NewValue(tftypes.Number, nil),
-			expectation: NewNumberNull(),
-		},
-		"wrongType": {
-			input:       tftypes.NewValue(tftypes.String, "oops"),
-			expectedErr: "can't unmarshal tftypes.String into *big.Float, expected *big.Float",
-		},
-	}
-	for name, test := range tests {
-		name, test := name, test
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			ctx := context.Background()
-
-			got, err := NumberType{}.ValueFromTerraform(ctx, test.input)
-			if err != nil {
-				if test.expectedErr == "" {
-					t.Errorf("Unexpected error: %s", err)
-					return
-				}
-				if test.expectedErr != err.Error() {
-					t.Errorf("Expected error to be %q, got %q", test.expectedErr, err.Error())
-					return
-				}
-				// we have an error, and it matches our
-				// expectations, we're good
-				return
-			}
-			if err == nil && test.expectedErr != "" {
-				t.Errorf("Expected error to be %q, didn't get an error", test.expectedErr)
-				return
-			}
-			if !got.Equal(test.expectation) {
-				t.Errorf("Expected %+v, got %+v", test.expectation, got)
-			}
-			if test.expectation.IsNull() != test.input.IsNull() {
-				t.Errorf("Expected null-ness match: expected %t, got %t", test.expectation.IsNull(), test.input.IsNull())
-			}
-			if test.expectation.IsUnknown() != !test.input.IsKnown() {
-				t.Errorf("Expected unknown-ness match: expected %t, got %t", test.expectation.IsUnknown(), !test.input.IsKnown())
-			}
-		})
-	}
-}
-
 func TestNumberValueToTerraformValue(t *testing.T) {
 	t.Parallel()
