+++ types/basetypes/string_value_test.go
@@ -9,69 +9,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-go/tftypes"
 )
 
-func TestStringTypeValueFromTerraform(t *testing.T) {
-	t.Parallel()
-
-	type testCase struct {
-		input       tftypes.Value
-		expectation attr.Value
-		expectedErr string
-	}
-	tests := map[string]testCase{
-		"true": {
-			input:       tftypes.NewValue(tftypes.String, "hello"),
-			expectation: NewStringValue("hello"),
-		},
-		"unknown": {
-			input:       tftypes.NewValue(tftypes.String, tftypes.UnknownValue),
-			expectation: NewStringUnknown(),
-		},
-		"null": {
-			input:       tftypes.NewValue(tftypes.String, nil),
-			expectation: NewStringNull(),
-		},
-		"wrongType": {
-			input:       tftypes.NewValue(tftypes.Number, 123),
-			expectedErr: "can't unmarshal tftypes.Number into *string, expected string",
-		},
-	}
-	for name, test := range tests {
-		name, test := name, test
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			ctx := context.Background()
-
-			got, err := StringType{}.ValueFromTerraform(ctx, test.input)
-			if err != nil {
-				if test.expectedErr == "" {
-					t.Errorf("Unexpected error: %s", err)
-					return
-				}
-				if test.expectedErr != err.Error() {
-					t.Errorf("Expected error to be %q, got %q", test.expectedErr, err.Error())
-					return
-				}
-				// we have an error, and it matches our
-				// expectations, we're good
-				return
-			}
-			if err == nil && test.expectedErr != "" {
-				t.Errorf("Expected error to be %q, didn't get an error", test.expectedErr)
-				return
-			}
-			if !got.Equal(test.expectation) {
-				t.Errorf("Expected %+v, got %+v", test.expectation, got)
-			}
-			if test.expectation.IsNull() != test.input.IsNull() {
-				t.Errorf("Expected null-ness match: expected %t, got %t", test.expectation.IsNull(), test.input.IsNull())
-			}
-			if test.expectation.IsUnknown() != !test.input.IsKnown() {
-				t.Errorf("Expected unknown-ness match: expected %t, got %t", test.expectation.IsUnknown(), !test.input.IsKnown())
-			}
-		})
-	}
-}
-
 func TestStringValueToTerraformValue(t *testing.T) {
 	t.Parallel()
